refactor: split texture things into multiple files
